+++ crates/rspack_binding_options/Cargo.toml
@@ -28,6 +28,7 @@ regex = { workspace = true }
 rspack_binding_macros = { path = "../rspack_binding_macros" }
 rspack_core = { path = "../rspack_core" }
 rspack_error = { path = "../rspack_error" }
+rspack_identifier = { path = "../rspack_identifier" }
 rspack_ids = { path = "../rspack_ids" }
 rspack_loader_sass = { path = "../rspack_loader_sass" }
 rspack_plugin_asset = { path = "../rspack_plugin_asset" }
