+++ crates/rspack_loader_runner_testing/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+  napi_build::setup();
+}
