--- crates/rspack_loader_runner_testing/src/content.rs
@@ -1,80 +1,0 @@
-use std::fmt::Debug;
-
-use rspack_error::{Error, Result};
-
-#[derive(Clone, PartialEq, Eq)]
-pub enum Content {
-  String(String),
-  Buffer(Vec<u8>),
-}
-
-impl Content {
-  pub fn try_into_string(self) -> Result<String> {
-    match self {
-      Content::String(s) => Ok(s),
-      Content::Buffer(b) => {
-        String::from_utf8(b).map_err(|e| rspack_error::internal_error!(e.to_string()))
-      }
-    }
-  }
-
-  pub fn as_bytes(&self) -> &[u8] {
-    match self {
-      Content::String(s) => s.as_bytes(),
-      Content::Buffer(b) => b,
-    }
-  }
-
-  pub fn into_bytes(self) -> Vec<u8> {
-    match self {
-      Content::String(s) => s.into_bytes(),
-      Content::Buffer(b) => b,
-    }
-  }
-}
-
-impl TryFrom<Content> for String {
-  type Error = Error;
-
-  fn try_from(content: Content) -> Result<Self> {
-    content.try_into_string()
-  }
-}
-
-impl From<Content> for Vec<u8> {
-  fn from(content: Content) -> Self {
-    content.into_bytes()
-  }
-}
-
-impl From<String> for Content {
-  fn from(s: String) -> Self {
-    Self::String(s)
-  }
-}
-
-impl From<Vec<u8>> for Content {
-  fn from(buf: Vec<u8>) -> Self {
-    Self::Buffer(buf)
-  }
-}
-
-impl Debug for Content {
-  fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    let mut content = f.debug_struct("Content");
-
-    let s = match self {
-      Self::String(s) => s.to_string(),
-      Self::Buffer(b) => String::from_utf8_lossy(b).to_string(),
-    };
-
-    let ty = match self {
-      Self::String(_) => "String",
-      Self::Buffer(_) => "Buffer",
-    };
-
-    content
-      .field(ty, &s[0..usize::min(s.len(), 20)].to_owned())
-      .finish()
-  }
-}
